feat: downgrade to prismic-reactjs
Using @slicemachine/init@1.0.2 to install this library causes `npm install`
to fail due to a peer dependency mismatch.

This happens because `@slicemachine/init@1.0.2` installs an old version
of `@prismicio/client` (v5, latest is v6).

As a temporary fix, this library is being downgraded from
`@prismicio/react` to `prismic-reactjs`. It will return to
`@prismicio/react` once a new version of `@slicemachine/init` is
published.